no message
git-svn-id: https://svn.grrrr.org/ext/trunk@377 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flatom_pr.cpp
@@ -81,7 +81,7 @@ bool flext::ScanAtom(t_atom &a,const char *buf)
 		break;
 #endif
 	case 1: // float
-		SetFloat(a,atof(tmp));
+		SetFloat(a,(float)atof(tmp));
 		break;
 	default: { // anything else is a symbol
 		char t = *c; *c = 0;
